flaskapp: adjust app for flask v1.0.2
flask v1.0.2 as provided by the python3-flask debian package in buster
does not automatically convert dict objects to JSON.

We switched to the debian package from the pypi flask v1.1.1 package
which supported automatic conversion of dict objects to JSON.

Backport the sample to make it compatible with the older flask version.
With this change the sample works on both versions.

Related-to: TOR-794

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>