tflife: update install method for TensorFlow Lite
Use apt to install TF Lite, as it is the main method presented on
https://www.tensorflow.org/lite/guide/python#install_tensorflow_lite_for_python
and it works well. Due to not being locked in on a specific version of
TF Lite, it also keeps this example updated over time.

If, for any reason, the user wants to install with PIP, they can
uncomment the respective section in the Dockerfile.

Related-to: DWC-2467
Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>