ci: split into dedicated CI files
Move the specific sample jobs to a split file, to help make the CI scripts
and jobs more readable, instead of keeping everything in a single big file.

Keep the shared scripts in a separate file. This file can be included in
all demos that need it.

Also, only trigger a specific sample jobs when something is changed in the
corresponding sample directory.

This, used together with "include:file" (https://docs.gitlab.com/ee/ci/yaml/#includefile)
allows to reuse the environemnt variables with versions (for example
DEBIAN_MAJOR, WESTON_PATCH) from the main project, instead of
duplicate it in projects that depend on it.

The parameter "strategy: depend" helps to avoid several jobs in
parallel, which seems to cause some failure, and also helps to visualize
if the pipeline really pass. See for more details the doc
https://docs.gitlab.com/ee/ci/yaml/#linking-pipelines-with-triggerstrategy

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>