ci: only run branch pipelines
Run only branch pipelines to avoid duplicated pipelines, as explained on
https://docs.gitlab.com/ee/ci/jobs/job_control.html#avoid-duplicate-pipelines
and use the template from https://docs.gitlab.com/ee/ci/yaml/index.html#workflowrules-templates
that has this exact purpose.

It avoids running MR pipelines, in turn prevents any of its
functionality. This can be easily changed in the future, if needed.

It also prevents MR pipelines to fail unless explicitly set both on the
parent and the child pipelines, as explained on
https://docs.gitlab.com/ee/ci/pipelines/parent_child_pipelines.html#merge-request-child-pipelines

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>