ci: add a final job that always run
All existing jobs only run if the respective demo directory has changes.
This makes a faster pipeline, but when no demo is changed then no
pipeline is run.

This is odd, for example, when editing the CI scripts or the README,
because a MR will keep "Checking pipeline status" forever.

Add a final job that always runs, even if no sample is changed, so the
MR process is sane.
It also provides a visual indication to the developer that the pipeline
has been triggered correctly.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>