sht31: use separate compose files for 32/64-bit
This helps to expose /dev devices more granularly e.g. when the
devices to be exposed are different for 32/64-bit.

We need to expose /dev/galcore and /dev/dri for 64 bit and 32 bit
devices respectively, for the kiosk container. /dev/galcore is only
available on i.MX8 and we do not want to expose whole /dev.

Related-to: TOR-2168

Signed-off-by: Abdur Rehman <abdur.rehman@toradex.com>