opencl: use multi-stage build
It is a best practice to use multi-stage to avoid dragging build
dependencies into the runtime container.

Use the Torizon SDK container to cross-compile the app, which is faster
than native emulation, and install the binary and runtime dependencies
in the runtime container.

This reduces the final container size from about 687MB to about 304MB.

Related-to: TOR-2169

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>