opencl: change the base image from wayland to debian
Make the default image for the sample as small as possible, by using
Debian base instead of Wayland base. It reduces the final image size
from about 304MB to 105MB.

This is useful because some customers may be only interesting in using
the GPU for general-purpose calculations on a headless system. One
example is running DL inference at the edge.

Leave the Wayland option as a comment, so users interested in a GUI can
very easily find out what is the preferred way to enable it.

p.s. the libwayland-client is still required, even if the entire Wayland
base isn't.

Related-to: TOR-2169

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>