opencl: add deploy step to the CI pipeline
The OpenCL sample has two programs: a system info tool and a benchmark.
These are great for providing a pre-built container, since they don't
require the users to change the source-code to evaluate the sample.
Furthermore, for a quick run or a quick test, having to build such
sample from source is a demotivator.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>