ci: attempt to get sources more than once
When configuring a local runner over a VPN connection, sometimes the Git
fetch of the repository fails with the error 'Failed to connect to
gitlab.int.toradex.com port 443'.

Attempting more than once is an easy solution that does not impact the
regular workflows - when the fetch succeeds at first try.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>