ci: enable use of torizon-containers' registry
We want to trigger the samples pipeline automatically via the
torizon-containers one. This will allow us to check if changes in the
containers are breaking samples.

Changes in .gitlab-ci.yml:

To do it, we need to pass the TORIZON_CONTAINERS_CI_REGISTRY when
this pipeline is triggered via torizon-containers, but we don't
otherwise (commom push, MR, etc.). So we add a rule to enforce this
behavior. Additionally, we set what folders trigger the pipeline when
changed. So each change self-contained to a sample, only triggers that
samples' pipeline.

Changes in .gitlab-ci-scripts.yml:

We also need to use the variable when building the samples' containers,
so we check to see if it's defined. If it is, we use it in the docker
build command.

Related-to: TOR-3038